retry tracer tests + configuration
